+++ src/impls/mod.rs
@@ -9,8 +9,6 @@ mod sets;
 
 #[cfg(feature = "alloc")]
 mod result;
-#[cfg(feature = "alloc")]
-mod vector;
 
 #[cfg(feature = "tuple")]
 mod tuples;