+++ tests/impl/main.rs
@@ -7,7 +7,6 @@ mod result;
 mod sets;
 mod try_unzip;
 mod tuples;
-mod vec;
 
 #[path = "../utils/mod.rs"]
 mod utils;